# docs/copilot-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Copilot Terraform Agent",
    "description": "Backend actions for Copilot Studio Terraform agent. Copilot Studio generates Terraform code; this API commits it to GitHub.",
    "version": "2.0.0",
    "contact": {
      "name": "Terraform Generator Team"
    }
  },
  "servers": [
    {
      "url": "https://copilot-terraform-agent.azurewebsites.net",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/copilot/generate": {
      "post": {
        "operationId": "commitTerraform",
        "summary": "Commit Terraform Code",
        "description": "Commit Terraform code (generated by Copilot Studio) to GitHub and create a Pull Request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["terraform_code"],
                "properties": {
                  "terraform_code": {
                    "type": "string",
                    "description": "The Terraform code generated by Copilot Studio to commit to GitHub",
                    "example": "terraform {\n  required_providers {\n    azurerm = {\n      source = \"hashicorp/azurerm\"\n    }\n  }\n}"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the infrastructure being created",
                    "example": "Create a storage account with blob container"
                  },
                  "location": {
                    "type": "string",
                    "description": "Azure region for the resources (for metadata)",
                    "default": "eastus",
                    "example": "eastus"
                  },
                  "create_pr": {
                    "type": "boolean",
                    "description": "Whether to create a GitHub Pull Request",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Terraform code committed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "pr_url": {
                      "type": "string",
                      "description": "URL of the created Pull Request",
                      "example": "https://github.com/owner/repo/pull/123"
                    },
                    "pr_number": {
                      "type": "integer",
                      "description": "PR number",
                      "example": 123
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable status message"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - terraform_code is required"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/copilot/status/{pr_number}": {
      "get": {
        "operationId": "getPRStatus",
        "summary": "Check PR Status",
        "description": "Check the status of a Terraform deployment Pull Request",
        "parameters": [
          {
            "name": "pr_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The PR number to check"
          }
        ],
        "responses": {
          "200": {
            "description": "PR status retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "pr_number": {
                      "type": "integer"
                    },
                    "pr_url": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string",
                      "enum": ["open", "closed", "merged"]
                    },
                    "merged": {
                      "type": "boolean"
                    },
                    "deployment_status": {
                      "type": "string",
                      "enum": ["pending_review", "deployed", "cancelled"]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "PR not found"
          }
        }
      }
    },
    "/api/templates": {
      "get": {
        "operationId": "listTemplates",
        "summary": "List Available Templates",
        "description": "Get a list of predefined Terraform templates",
        "responses": {
          "200": {
            "description": "Templates list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "templates": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "prompt": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "operationId": "healthCheck",
        "summary": "Health Check",
        "description": "Check if the service is running",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "healthy"
                    },
                    "service": {
                      "type": "string",
                      "example": "Copilot Terraform Agent"
                    },
                    "version": {
                      "type": "string",
                      "example": "1.0.0"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}